Split event.c into per-API files
