+++ src/linux/meson.build
@@ -1,12 +1,14 @@
 check_headers += [
   'linux/if_tun.h',
   'netpacket/packet.h',
-  'sys/epoll.h',
 ]
 
 check_functions += 'recvmmsg'
 
-src_tincd += files('device.c')
+src_tincd += files(
+  'device.c',
+  'event.c',
+)
 
 dep_libsystemd = dependency('libsystemd', required: opt_systemd)
 if dep_libsystemd.found()
