+++ src/solaris/meson.build
@@ -1,4 +1,7 @@
-src_tincd += files('device.c')
+src_tincd += [
+  files('device.c'),
+  src_event_select,
+]
 
 deps_common += cc.find_library('libsocket')
